Module stat_summoner::module::randomchampions::utils

Functions

get_champions_by_role ğŸ”’

âš™ï¸ Retrieves all champions that match the specified role from the database.

get_champions_with_no_role ğŸ”’

âš™ï¸ Retrieves all champions from the database without filtering by role.

get_list_champions

âš™ï¸ Retrieves a list of champions from the database, optionally filtered by role.

get_random_champion

âš™ï¸ Selects a random champion from a list of ChampionData.

match_role_with_database_roles ğŸ”’

âš™ï¸ Maps a Role enum value to its corresponding string representation as stored in the database.