stat_summoner::module::suggestions::utils
Function is_user_blacklisted

pub async fn is_user_blacklisted(
    collection: &Collection<User>,
    user_id: &str,
) -> Result<bool>

⚙️ Checks if a user is blacklisted.

This asynchronous function queries a MongoDB collection to check whether a user with the specified user_id exists in the blacklist. It returns true if the user is found, and false otherwise.

Parameters:

collection: A reference to a MongoDB collection of BlackList documents.

user_id: A string slice representing the ID of the user to check.

Returns:

Result<bool, mongodb::error::Error>: On success, returns true if the user is blacklisted, false otherwise. On failure, returns an Error from MongoDB detailing the issue.

Example:

let is_blacklisted = is_user_blacklisted(&collection, "123456789012345678").await?;
if is_blacklisted {
    println!("User is blacklisted.");
} else {
    println!("User is not blacklisted.");
}

⚠️ Notes:

The function uses MongoDB’s find_one method to search for a document with a matching user_id.

If no matching document is found, the function returns false.

Related Structures:

User: Represents the structure of a User in the database.

Dependencies:

Requires a MongoDB collection containing BlackList documents.